Refactor modules in src
