+++ dev.html
@@ -66,7 +66,7 @@
     </template>
     <!-- Chargement des modules en ES modules -->
     <script src="env-loader.js"></script>
-    <script type="module" src="calendar.js"></script>
+    <script type="module" src="src/calendar.js"></script>
     <script type="module" src="test-api.js"></script>
 </body>
 </html>
