+++ index.html
@@ -66,6 +66,6 @@
     </template>
     <!-- Chargement des modules en ES modules -->
     <script src="env-loader.js"></script>
-    <script type="module" src="calendar.js"></script>
+    <script type="module" src="src/calendar.js"></script>
 </body>
 </html>
